docs/requirements.txt
sphinx==4.5.0
sphinx-rtd-theme==1.0.0
sphinx-copybutton==0.5.0
myst-parser==0.16.1
docutils==0.17
sphinx-tabs
pyembed-rst